==> 1_first_cleaning.ipynb <==
import pandas as pd
import pickle
from functools import reduce

#a=pd.read_csv('genomas_eliminados.csv',header=None,names=['code'])

a=pd.read_csv('Lista_final_editada_20190128.csv',sep=';',header=None)

b=a.iloc[:,0].apply(lambda x: x[0:-8])
bx=b.tolist()

a

with open('rhodo_contig_genome_dict.pkl','rb') as f:
    values=pickle.load(f)

f=lambda x:x[0:-13]

d2 = {f(k):v for k, v in values.items()}

q=reduce(lambda x,y: x+y,[d2[i] for i in bx])

values

nodos=pd.read_csv('nodes_all_modified.csv')

nodos['iden']=nodos['Id'].apply(lambda x:x.split('.')[0]+'.1')

#nodos.set_index('iden',inplace=True)

nolim=nodos[~(nodos['iden'].apply(lambda x:values[x]).isin(bx))]

exclu=nodos[(nodos['iden'].apply(lambda x:values[x]).isin(bx))]

print(len(nolim))
print(len(exclu))
print(len(nodos))

nolim2=nolim.copy().drop('iden',axis=1)

gen=pd.unique(nolim['iden'].apply(lambda x:values[x]))

len(gen)

#nolim2.to_csv('v2_nodes_after_cleaning.csv',index=False)

yu=pd.read_csv('edges_06_all.csv')

yu2=yu.loc[(yu['Source'].isin(nolim['Id'])) & (yu['Target'].isin(nolim['Id']))]

yu3=yu2[yu2['Source']!=yu2['Target']]

rt=pd.unique(yu3['Source'].append(yu3['Target']))

len(rt)

#yu3.to_csv('v2_edges_06_after_cleaning.csv',index=False)
