==> 2_filtering_by_bgc_features.ipynb <==
from Bio import SeqIO
import os
import re
import pandas as pd
from functools import reduce
from shutil import copyfile
#import community
import pickle

pattern=re.compile('\((.*?)\)')

def check(secu,tipo):
    for i in secu.features:
        if i.type=='cluster':
            if i.qualifiers['product'][0]==tipo:
                return True
            else:
                return False
            break

yo=pd.read_csv('v2_nodes_after_cleaning.csv')

yo2=yo.copy()

ty=[]
for w in yo2['Id']:
    boo=SeqIO.read('clusters/'+w+'.gbk','genbank')
    ty.append(len(boo))

yo3=yo2.copy()
yo3['length']=ty

yo3=yo3[['Id','length']].sort_values('length')

clu=[]
leng=[]
cds=[]
bio_cds=[]
counter1=0
counter2=0
#amp=[]
val_amp=0
for i,l in zip(yo3['Id'],yo3['length']):
    q=SeqIO.read('clusters/'+i+'.gbk','genbank')
    sumario_cds=[cat.qualifiers for cat in q.features if (cat.type=='CDS')]
    o_sec=[j for j in sumario_cds if ('sec_met' in j.keys())]
    nc=[]
    for a in o_sec:
        tipox=a['sec_met'][0].split(':')[1].strip()
        try:
            kind=a['sec_met'][2].split(':')[1].strip()
        except:
            kind='none'
        if check(q,'nrps'):
            if (kind=='biosynthetic')&(tipox=='nrps'):
                nc.append(a)
        else:
            if kind=='biosynthetic':
                nc.append(a)
    sumario=[cat.qualifiers for cat in q.features if (cat.type=='PFAM_domain')]
    f=lambda x:x['domain'][0]
    if (len(sumario)!=0)&(check(q,'nrps')):
        dominios=list(map(f,sumario))
        val_amp=dominios.count('AMP-binding')
    #if (len(sumario_cds)>=2) & (len(nc)>=1) & (len(sumario_cds)>len(nc)) & (not(check(q,'nrps'))):
    #if (check(q,'nrps'))&(len(sumario_cds)>=2) & (len(nc)>=1) & (len(sumario_cds)>len(nc)):
        #counter1+=1
    #if (check(q,'nrps'))&(len(sumario_cds)>=2) & (len(nc)>=1) & (len(sumario_cds)>len(nc)) & (val_amp>=2):
        #counter2+=1
    if (check(q,'nrps'))&(len(sumario_cds)>=2) & (len(nc)>=1) & (len(sumario_cds)>len(nc)) & (val_amp>=2):
        clu.append(i)
        leng.append(l)
        cds.append(len(sumario_cds))
        bio_cds.append(len(nc))
        #amp.append(val_amp)
    if (len(sumario_cds)>=2) & (len(nc)>=1) & (len(sumario_cds)>len(nc)) & (not(check(q,'nrps'))):
        clu.append(i)
        leng.append(l)
        cds.append(len(sumario_cds))
        bio_cds.append(len(nc))

counter2

tempo=pd.DataFrame({'Id':clu,'length':leng,'num_cds':cds,'num_bio_cds':bio_cds}) #,'amp_binding':amp

tu1=yo.copy()
spe=yo[['Id','reduccion']]

tu1.set_index('Id',inplace=True)

selec=tu1.loc[tempo['Id'].tolist(),:]

selec2=selec.reset_index('Id')

with open('rhodo_contig_genome_dict.pkl','rb') as f:
    values=pickle.load(f)

tre=selec2['Id'].apply(lambda x:x[0:-11].split('.')[0]+'.1').apply(lambda x:values[x])

len(pd.unique(tre))

yo['reduccion'].value_counts()

#selec2[~selec2['Label'].isna()]
selec2['reduccion'].value_counts()

#selec2['reduccion'].value_counts().to_excel('conteo_154genomas+BGC_check_control.xlsx')

selec2.to_csv('v2_nodes_after_cleaning_checkcontrol.csv',index=False)

chosen=selec2['Id'].tolist()

tempo2=tempo.merge(spe,how='left',on='Id')

tempo2=tempo2.sort_values('length')

tempo2.to_csv('v2_info_nodes_after_cleaning_checkcontrol.csv',index=False)

edgo=pd.read_csv('v2_edges_06_after_cleaning.csv')

edgonuevo=edgo.loc[(edgo['Source'].isin(chosen))&(edgo['Target'].isin(chosen))]

edgonuevo.to_csv('v2_edges_06_after_cleaning_checkcontrol.csv',index=False)
